+++ tests/test_hosts.py
@@ -11,7 +11,8 @@ class DependencyManagerTestCase(unittest.TestCase):
         # Basic tests
         m_context = {
             'STATIC_URL': '/static/',
-            'echarts_version': '3.7.0'
+            'echarts_version': '3.7.0',
+            'baidu_map_ak': 'foo23zoo'
         }
         manager = DependencyManager.create_default(
             context=m_context,
@@ -23,6 +24,8 @@ class DependencyManagerTestCase(unittest.TestCase):
             'https://cdn.bootcss.com/echarts/3.7.0/echarts.min.js',
             manager.resolve_url('echarts.min')
         )
+        self.assertEqual('https://api.map.baidu.com/api?v=2.0&ak=foo23zoo',
+                         manager.resolve_url('https://api.map.baidu.com/api?v=2.0&ak=xxx'))
 
     def test_map_host(self):
         m_context = {